requirements.txt
Flask
Flask-SQLAlchemy
Flask-Bcrypt
Flask-Login
Flask-WTF
python-dotenv
mira_sdk 
